+++ muna/beta/metadata/__init__.py
@@ -6,6 +6,7 @@
 from ._ort import OnnxRuntimeExecutionProvider, OnnxRuntimeOptimizationLevel
 from ._torch import KVCacheConfig, TorchExporter
 from .coreml import CoreMLInferenceMetadata
+from .cuda import CudaArchitecture, GeneratedCudaKernelInferenceMetadata
 from .executorch import ExecuTorchInferenceBackend, ExecuTorchInferenceMetadata
 from .iree import IREEInferenceBackend, IREEInferenceMetadata
 from .litert import LiteRTInferenceMetadata
@@ -15,10 +16,7 @@ from .onnxruntime import OnnxRuntimeInferenceMetadata, OnnxRuntimeInferenceSessi
 from .openvino import OpenVINOInferenceMetadata
 from .qnn import QnnInferenceBackend, QnnInferenceMetadata, QnnInferenceQuantization
 from .sglang import SGLangInferenceMetadata
-from .tensorrt import (
-    CudaArchitecture, TensorRTInferenceMetadata, TensorRTInferenceSessionMetadata,
-    TensorRTHardwareCompatibility
-)
+from .tensorrt import TensorRTInferenceMetadata, TensorRTInferenceSessionMetadata
 from .tensorrt_llm import (
     TensorRTLLMInferenceMetadata, TensorRTLLMQuantization,
     TensorRTLLMKVCacheQuantization
@@ -28,22 +26,23 @@ from .tflite import TFLiteInterpreterMetadata
 from .vllm import vLLMInferenceMetadata
 
 InferenceMetadata = (
-    CoreMLInferenceMetadata             |
-    ExecuTorchInferenceMetadata         |
-    IREEInferenceMetadata               |
-    LiteRTInferenceMetadata             |
-    LlamaCppInferenceMetadata           |
-    MLXInferenceMetadata                |
-    MLXInferenceSessionMetadata         |
-    OnnxRuntimeInferenceMetadata        |
-    OnnxRuntimeInferenceSessionMetadata |
-    OpenVINOInferenceMetadata           |
-    QnnInferenceMetadata                |
-    SGLangInferenceMetadata             |
-    TensorRTInferenceMetadata           |
-    TensorRTInferenceSessionMetadata    |
-    TensorRTLLMInferenceMetadata        |
-    TensorRTRTXInferenceMetadata        |
-    TFLiteInterpreterMetadata           |
+    CoreMLInferenceMetadata                 |
+    ExecuTorchInferenceMetadata             |
+    GeneratedCudaKernelInferenceMetadata    |
+    IREEInferenceMetadata                   |
+    LiteRTInferenceMetadata                 |
+    LlamaCppInferenceMetadata               |
+    MLXInferenceMetadata                    |
+    MLXInferenceSessionMetadata             |
+    OnnxRuntimeInferenceMetadata            |
+    OnnxRuntimeInferenceSessionMetadata     |
+    OpenVINOInferenceMetadata               |
+    QnnInferenceMetadata                    |
+    SGLangInferenceMetadata                 |
+    TensorRTInferenceMetadata               |
+    TensorRTInferenceSessionMetadata        |
+    TensorRTLLMInferenceMetadata            |
+    TensorRTRTXInferenceMetadata            |
+    TFLiteInterpreterMetadata               |
     vLLMInferenceMetadata
 )